requirements.txt
Django==1.6
django-ibge